requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
altgraph==0.17.4
annotated-types==0.7.0
attrs==23.2.0
blinker==1.8.2
Brotli==1.1.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
ConfigArgParse==1.7
exceptiongroup==1.2.1
Faker==25.8.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Login==0.6.3
gevent==24.2.1
geventhttpclient==2.3.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
locust==2.29.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgpack==1.0.8
mysql-connector-python==8.4.0
outcome==1.3.0.post0
packaging==24.1
pefile==2023.2.7
pillow==10.3.0
pluggy==1.5.0
psutil==6.0.0
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pyinstaller==6.8.0
pyinstaller-hooks-contrib==2024.7
PySocks==1.7.1
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pywin32-ctypes==0.2.2
pyzmq==26.0.3
reloadium==1.5.1
requests==2.32.3
rich==13.7.1
selenium==4.22.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tk==0.1.0
tomli==2.0.1
trio==0.26.0
trio-websocket==0.11.1
ttkbootstrap==1.10.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
Werkzeug==3.0.3
wsproto==1.2.0
zope.event==5.0
zope.interface==6.4.post2
